+++ pkg/coinbase/validators_test.go
@@ -131,6 +131,58 @@ func (s *ValidatorSuite) TestGetValidator_Failure() {
 	s.EqualError(err, "APIError{HttpStatusCode: 500, Code: unknown, Message: some error calling api}")
 }
 
+func (s *ValidatorSuite) TestGetters() {
+	validator := NewValidator(api.Validator{
+		ValidatorId: "validator-1",
+		NetworkId:   EthereumHolesky,
+		AssetId:     Eth,
+		Status:      api.VALIDATORSTATUS_ACTIVE,
+		Details: &api.ValidatorDetails{
+			EthereumValidatorMetadata: &api.EthereumValidatorMetadata{
+				Index:             "0",
+				PublicKey:         "public-key-1",
+				WithdrawalAddress: "withdrawal-address-1",
+				Slashed:           false,
+				ActivationEpoch:   "epoch-1",
+				ExitEpoch:         "exit-epoch-2",
+				WithdrawableEpoch: "withdrawable-epoch-3",
+				Balance: api.Balance{
+					Amount: "100",
+					Asset: api.Asset{
+						NetworkId: EthereumHolesky,
+						AssetId:   Eth,
+					},
+				},
+				EffectiveBalance: api.Balance{
+					Amount: "200",
+					Asset: api.Asset{
+						NetworkId: EthereumHolesky,
+						AssetId:   Eth,
+					},
+				},
+			},
+		},
+	})
+
+	s.Assert().Equal("validator-1", validator.ID())
+	s.Assert().Equal(EthereumHolesky, validator.NetworkID())
+	s.Assert().Equal(Eth, validator.AssetID())
+	s.Assert().Equal(api.VALIDATORSTATUS_ACTIVE, validator.Status())
+	s.Assert().Equal("0", validator.Index())
+	s.Assert().Equal("public-key-1", validator.PublicKey())
+	s.Assert().Equal("withdrawal-address-1", validator.WithdrawalAddress())
+	s.Assert().Equal(false, validator.Slashed())
+	s.Assert().Equal("epoch-1", validator.ActivationEpoch())
+	s.Assert().Equal("exit-epoch-2", validator.ExitEpoch())
+	s.Assert().Equal("withdrawable-epoch-3", validator.WithdrawableEpoch())
+	s.Assert().Equal("100", validator.Balance().Amount)
+	s.Assert().Equal(EthereumHolesky, validator.Balance().Asset.NetworkId)
+	s.Assert().Equal(Eth, validator.Balance().Asset.AssetId)
+	s.Assert().Equal("200", validator.EffectiveBalance().Amount)
+	s.Assert().Equal(EthereumHolesky, validator.EffectiveBalance().Asset.NetworkId)
+	s.Assert().Equal(Eth, validator.EffectiveBalance().Asset.AssetId)
+}
+
 func (s *ValidatorSuite) mockSuccessfulListValidators(ctx context.Context, networkId string, assetId string, mockValidators *api.ValidatorList) {
 	s.T().Helper()